Add Windows compilation compatibility fixes
- Add MSVC compatibility macros for __asm__ and __volatile__ in CUDA kernels
- Fix long/int64_t type mismatches for Windows compatibility
- Add platform-specific compiler flags in setup.py for Windows/MSVC support
- Fix uint64_t type issues in hash functions for Windows
- Enables native Windows compilation without requiring WSL2

Fixes compilation errors:
- 'identifier __asm__ is undefined' in CUDA kernels
- 'argument of type long * is incompatible with parameter of type int64_t *'
- uint64_t type compatibility issues on Windows

+++ torchsparse/backend/hash/hash_cuda.cu
@@ -11,7 +11,7 @@ __global__ void hash_kernel(int N, const int *__restrict__ data,
   int i = blockDim.x * blockIdx.x + threadIdx.x;
   if (i < N) {
     data += i * 4;
-    uint64_t hash = 14695981039346656037UL;
+    unsigned long long hash = 14695981039346656037UL;
     for (int j = 0; j < 4; j++) {
       hash ^= (unsigned int)data[j];
       hash *= 1099511628211UL;
@@ -43,7 +43,7 @@ __global__ void kernel_hash_kernel(int N, int K, const int *__restrict__ data,
       cur_coord[j] = data[j] + kernel_offset[k * 3 + j - 1];
     }
     cur_coord[0] = data[0];
-    uint64_t hash = 14695981039346656037UL;
+    unsigned long long hash = 14695981039346656037UL;
     for (int j = 0; j < 4; j++) {
       hash ^= (unsigned int)cur_coord[j];
       hash *= 1099511628211UL;
